# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Finanzas App API
  version: 1.0.0
  description: API v1 para app financiera Chile-first (CLP).
servers:
  - url: https://api.finanzasapp.cl/v1
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Registro manual
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login manual
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Sesion creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/google:
    post:
      tags: [Auth]
      summary: Login Google con token id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idToken]
              properties:
                idToken:
                  type: string
      responses:
        '200':
          description: Sesion creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refrescar access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        fullName: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
            fullName: { type: string }
    RefreshResponse:
      type: object
      properties:
        accessToken: { type: string }
...